Change preprocessing command to compilation
Instead of outputting the preprocessor result (which I assumed would be
faster), instead just compile the file.  This results in a huge
improvement to Visual Studio times since it was outputting hundreds of
thousands of newlines when being passed `/E`.

Additionally we can now define `struct foo {};` and get compilation
errors if the guard actually isn't working correctly.  This caught a few
bugs.

Update the README.md with the new results.